Update Python template
Header text in a Python file should be contained within a Python docstring (the unquoted text will return an error).

The requests module is a  third-party import.

The header text should probably list required third-party packages for the script to function - unless these are documented elsewhere in the code (like a requirements.txt file or a Pipfile, etc.).

Constants (Python variables that should not change throughout the script's execution should be in ALL CAPS).

It should be standard practice to performa a simple response code check to verify if the request succeeded or not.  This helps provide more accurate and useful errors to the users when a response fails.  Without it, they don't get an error until they try access, parse or use the data.

It should also be a standard practice to parse the data once we have verified that the request was successful.

Unformatted JSON output can be scarry, formatted JSON is a little less scarry.  :100: